Update README.md
In this paper, we analyzed brain activity data related to right and left hand movements in the beta frequency range (β). The aim of the study was to investigate the spatiotemporal dynamics of brain activity in the interval from -800 ms to +1600 ms relative to the stimulus/event. To visualize the results, topographic maps were constructed reflecting the distribution of activity over the surface of the head.